+++ lib/puppet-debugger/support/input_responders.rb
@@ -3,7 +3,7 @@ module PuppetDebugger
   module Support
     module InputResponders
       def static_responder_list
-        %w(exit functions classification vars facterdb_filter krt facts
+        %w(exit functions classification vars facterdb_filter krt facts types
            resources classes whereami datatypes play reset help)
       end
 
@@ -15,7 +15,8 @@ module PuppetDebugger
 
       # @source_file and @source_line_num instance variables must be set for this
       # method to show the surrounding code
-      # @return [String] - string output of the code surrounded by the breakpoint or nil if file or line_num do not exist
+      # @return [String] - string output of the code surrounded by the breakpoint or nil if file
+      # or line_num do not exist
       def whereami(_command = nil, _args = nil)
         file = @source_file
         line_num = @source_line_num
@@ -26,7 +27,29 @@ module PuppetDebugger
           else
             code = DebuggerCode.from_file(file, :puppet)
           end
-          return code.with_marker(line_num).around(line_num, 5).with_line_numbers.with_indentation(5).with_file_reference.to_s
+          return code.with_marker(line_num).around(line_num, 5)
+                     .with_line_numbers.with_indentation(5).with_file_reference.to_s
+        end
+      end
+
+      # @return - returns a list of types available to the environment
+      # if a error occurs we we run the types function again
+      def types(_args = [])
+        loaded_types = []
+        begin
+          # this loads all the types, if already loaded the file is skipped
+          Puppet::Type.loadall
+          Puppet::Type.eachtype do |t|
+            next if t.name == :component
+            loaded_types << t.name.to_s
+          end
+          loaded_types.ai
+        rescue Puppet::Error => e
+          puts e.message.red
+          Puppet.info(e.message)
+          # prevent more than two calls and recursive loop
+          return if caller_locations(1,10).find_all{|f| f.label == 'types' }.count > 2
+          types
         end
       end